@@ -42,7 +42,7 @@
     "@types/sanitize-html": "^2.6.0",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
-    "eslint": "^8.4.1",
+    "eslint": "^8.5.0",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-react-hooks": "^4.3.0",
@@ -56,23 +56,20 @@
     "@cuvva/ksuid": "^1.0.2",
     "@monaco-editor/react": "^4.3.1",
     "@sentry/cli": "^1.71.0",
-    "@sentry/electron": "^2.5.3",
+    "@sentry/electron": "^2.5.4",
     "@sentry/react": "^6.16.1",
     "@sentry/tracing": "^6.16.1",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-router": "^5.1.16",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
     "@types/url-parse": "^1.4.5",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "chokidar": "^3.5.2",
-    "connected-react-router": "^6.9.2",
     "cross-env": "^7.0.3",
     "keytar": "^7.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
     "tslog": "^3.3.0",
     "url-parse": "^1.5.3",
     "valid-filename": "^4.0.0",
